Development

+++ routes/chapterRoute.js
@@ -6,11 +6,12 @@ const ContentVideo = require("../models/contnetVideoModel")
 const Course = require("../models/courseModel")
 const router = express.Router()
 const {HTTPStatusCode,ErrorMessages} = require("../global.ts")
-
+const {authenticateToken} = require("../authenticateToken")
+const Instructor = require("../models/instructorModel");
 //Create
-router.post('/chapter/create' , async(req,res) =>{
-    let {chapterName,courseId} = req.body
-   
+router.post('/chapter/create' ,authenticateToken, async(req,res) =>{
+    let {chapterName,course} = req.body
+    const userid = req.user.id;
     try {
         // const allCourse = await Course.aggregate([
         //     {
@@ -21,24 +22,27 @@ router.post('/chapter/create' , async(req,res) =>{
         // ]);
        
         //Course Id Valid or not
-        const courseID = await Course.findById({_id:courseId})
-        const chapterExist = await Chapter.findOne({chapterName:chapterName})
-        if(chapterExist){
-            return res
-                .status(HTTPStatusCode.BAD_REQUEST)
-                .json({message:ErrorMessages.CHAPTER_EXIST})
-        }
-
-        const createChapter = await Chapter.create({
-            chapterName:chapterName,
-            course:courseID
-        })
-        return res
-            .status(HTTPStatusCode.CREATED)
-            .json({
-                message:ErrorMessages.CREATED,
-                data:createChapter
+        const instructorExist = await Instructor.findById({_id:userid})
+        if(instructorExist){
+            const courseID = await Course.findById({_id:course})
+            const chapterExist = await Chapter.findOne({chapterName:chapterName})
+            if(chapterExist){
+                return res
+                    .status(HTTPStatusCode.BAD_REQUEST)
+                    .json({message:ErrorMessages.CHAPTER_EXIST})
+            }
+            const createChapter = await Chapter.create({
+                chapterName:chapterName,
+                course:courseID,
+                createdBy:instructorExist
             })
+            return res
+                .status(HTTPStatusCode.CREATED)
+                .json({
+                    message:ErrorMessages.CREATED,
+                    data:createChapter
+                })
+        }
     } catch (error) {
         return res
             .status(HTTPStatusCode.INTERNAL_SERVER)
@@ -49,7 +53,7 @@ router.post('/chapter/create' , async(req,res) =>{
 })
 
 //Get All Chapter
-router.get('/getAllChapter' , async(req,res) =>{
+router.get('/getAllChapter' , authenticateToken,async(req,res) =>{
     try{
         const getAllChapter = await Chapter.find()
         
@@ -80,6 +84,7 @@ router.get('/getAllChapter' , async(req,res) =>{
 //Get Single Contnent
 router.get('/chapter/:id', async(req,res) =>{
     const id = req.params.id;
+ 
     try {
         if(ObjectId.isValid(id)){
             const chapterdetails = await Chapter.findOne({_id:id})
@@ -118,35 +123,40 @@ router.get('/chapter/:id', async(req,res) =>{
     
 })
 //Update
-router.patch('/chapter/update/:id', async(req,res) => {
+router.patch('/chapter/update/:id',authenticateToken, async(req,res) => {
     const id = req.params.id;
+    const instructorid = req.user.id;
+    const instructorExist = await Instructor.findById({_id:instructorid})
     try {
-        if(ObjectId.isValid(id)){
-            const ChapterUpdate = await Chapter.findByIdAndUpdate(id, req.body, {
-                new:true
-            })
-
-            if(ChapterUpdate){
-                return res
-                .status(HTTPStatusCode.OK)
-                .json({message: ErrorMessages.UPDATED,
-                    data :ChapterUpdate
+        if(instructorExist){
+            if(ObjectId.isValid(id)){
+                const ChapterUpdate = await Chapter.findByIdAndUpdate(id, req.body, {
+                    new:true
                 })
+    
+                if(ChapterUpdate){
+                    return res
+                    .status(HTTPStatusCode.OK)
+                    .json({message: ErrorMessages.UPDATED,
+                        data :ChapterUpdate
+                    })
+                }
+                else{
+                    return res
+                    .status(HTTPStatusCode.BAD_REQUEST)
+                    .json({message:ErrorMessages.NOT_EXISTS})
+               }
+              
             }
             else{
                 return res
-                .status(HTTPStatusCode.BAD_REQUEST)
-                .json({message:ErrorMessages.NOT_EXISTS})
-           }
-          
+                    .status(HTTPStatusCode. BAD_REQUEST)
+                    .json({
+                        message: ErrorMessages. WRONG_CREDENTIALS
+                    }) 
+            } 
         }
-        else{
-            return res
-                .status(HTTPStatusCode. BAD_REQUEST)
-                .json({
-                    message: ErrorMessages. WRONG_CREDENTIALS
-                }) 
-        }  
+ 
     }catch(error){
         return res
             .status(HTTPStatusCode.INTERNAL_SERVER)
@@ -158,9 +168,12 @@ router.patch('/chapter/update/:id', async(req,res) => {
 })
 
 //Delete
-router.delete('/chapter/delete/:id', async(req,res) => {
+router.delete('/chapter/delete/:id', authenticateToken,async(req,res) => {
     const id = req.params.id;
-    try{
+    const instructorid = req.user.id;
+    const instructorExist = await Instructor.findById({_id:instructorid})
+    try {
+        if(instructorExist){
             if(ObjectId.isValid(id)){
                 const ChapterDelete = await Chapter.findByIdAndDelete({_id:id})
                 if(ChapterDelete){
@@ -186,6 +199,7 @@ router.delete('/chapter/delete/:id', async(req,res) => {
                         message:ErrorMessages.WRONG_CREDENTIALS
                     }) 
             }
+        }
     }
     catch(error){
         return res
