+++ routes/courseRoute.js
@@ -1,6 +1,7 @@
 const express = require("express")
 const mongoose = require("mongoose")
 const router  = express.Router();
+const Instructor = require("../models/instructorModel");
 const Course = require("../models/courseModel")
 const Category = require("../models/categoryModel")
 const Subcategory = require("../models/subcategoryModel")
@@ -8,9 +9,11 @@ const ProgrammingLanguage = require("../models/programmingLanguagesModel")
 const Language = require("../models/languageModel")
 const ObjectId = require("mongoose").Types.ObjectId
 const  {HTTPStatusCode,ErrorMessages} =  require("../global.ts")
-console.log(ErrorMessages. USER_EXIST);
+const {authenticateToken} = require("../authenticateToken")
+const jwt = require("jsonwebtoken")
+// console.log(ErrorMessages. USER_EXIST);
 
-router.post('/course/createCourse', async(req,res) =>{
+router.post('/course/createCourse',authenticateToken,async(req,res) =>{
     const { 
             name,
             categoryId,
@@ -29,43 +32,60 @@ router.post('/course/createCourse', async(req,res) =>{
             file,
             review,
             instructor,
-            video
+            video,
+            createdBy
         } = req.body;
-    try {
+        const userid = req.user.id;
+        console.log(userid);
+        try{
+            const category = await Category.findById({_id:categoryId});
+            const subCategory = await Subcategory.findById({_id:subCategoryId});
+            const language = await Language.findById({_id:languageId});
+            const programmingLanguage = await ProgrammingLanguage.findById({_id:programmingLanguageId})
+            const courseExist = await Course.findOne({name:name})
+            const instructorExist = await Instructor.findById({_id:userid})
+            if(instructorExist){
+                
+                if(courseExist){
+                    return res
+                        .status(HTTPStatusCode.BAD_REQUEST)
+                        .json({
+                            message: ErrorMessages.COURSE_EXIST,
+                    })
+                }
+            
+                const courseCreate = await Course.create({
+                    //Model Name : Value of that Field
+                    name:name,
+                    overview:overview,
+                    description:description,
+                    requirement:requirement,
+                    category:category,
+                    subCategory:subCategory,
+                    programmingLanguage:programmingLanguage,
+                    language:language,
+                    price:price,
+                    discount:discount,
+                    level:level,
+                    deadline:deadline,
+                    createdBy:instructorExist
+                }) 
+                return res
+                        .status(HTTPStatusCode.CREATED)
+                        .json({
+                            message: ErrorMessages.CREATED,
+                            data:courseCreate
 
-        const category = await Category.findById({_id:categoryId});
-        const subCategory = await Subcategory.findById({_id:subCategoryId});
-        const language = await Language.findById({_id:languageId});
-        const programmingLanguage = await ProgrammingLanguage.findById({_id:programmingLanguageId})
-        const courseExist = await Course.findOne({name:name})
-        if(courseExist){
-            return res
-                .status(HTTPStatusCode.BAD_REQUEST)
-                .json({
-                    message: ErrorMessages.COURSE_EXIST,
-            })
-        }
-        const courseCreate = await Course.create({
-            //Model Name : Value of that Field
-            name:name,
-            overview:overview,
-            description:description,
-            requirement:requirement,
-            category:category,
-            subCategory:subCategory,
-            programmingLanguage:programmingLanguage,
-            language:language,
-            price:price,
-            discount:discount,
-            level:level,
-            deadline:deadline
-        }) 
-        return res
-                .status(HTTPStatusCode.CREATED)
-                .json({
-                    message: ErrorMessages.CREATED,
-                    data: courseCreate
-                })
+                        })
+            }else {
+                return res
+                    .status(HTTPStatusCode.BAD_REQUEST)
+                    .json({
+                            message:ErrorMessages.EMAIL_DOES_NOT_EXIST,
+                            
+                    })
+            }
+
     } catch (error) {
         return res
             .status(HTTPStatusCode.INTERNAL_SERVER)
@@ -75,33 +95,39 @@ router.post('/course/createCourse', async(req,res) =>{
     }
 })
 //Get All Data
-router.get('/getAllCourse', async(req,res) =>{
+router.get('/getAllCourse',authenticateToken, async(req,res) =>{
+    const userid = req.user.id;
+    console.log(userid);
     try {
-        const getAllCourse = await Course.find();
-        for (const fieldNames of getAllCourse) {
-            const category = await Category.findById({_id:fieldNames.category})
-            if(category){
-                fieldNames.category = category
+        const instructorExist = await Instructor.findById({_id:userid})
+        if(instructorExist){
+            console.log(instructorExist);
+            const getAllCourse = await Course.find({createdBy: userid});
+            for (const fieldNames of getAllCourse) {
+                const category = await Category.findById({_id:fieldNames.category})
+                if(category){
+                    fieldNames.category = category
+                }
+                const subCategory = await Subcategory.findById({_id:fieldNames.subCategory})
+                if(subCategory){
+                    fieldNames.subCategory = subCategory
+                }
+                const programmingLanguage = await ProgrammingLanguage.findById({_id:fieldNames.programmingLanguage})
+                if(programmingLanguage){
+                    fieldNames.programmingLanguage = programmingLanguage
+                }
+                const language = await Language.findById({_id:fieldNames.language});
+                if(language){
+                    fieldNames.language = language
+                }
             }
-            const subCategory = await Subcategory.findById({_id:fieldNames.subCategory})
-            if(subCategory){
-                fieldNames.subCategory = subCategory
-            }
-            const programmingLanguage = await ProgrammingLanguage.findById({_id:fieldNames.programmingLanguage})
-            if(programmingLanguage){
-                fieldNames.programmingLanguage = programmingLanguage
-            }
-            const language = await Language.findById({_id:fieldNames.language});
-            if(language){
-                fieldNames.language = language
-            }
+            return res
+                .status(HTTPStatusCode.OK)
+                .json({
+                    message: ErrorMessages.GETDATA,
+                    data: getAllCourse
+                })
         }
-        return res
-            .status(HTTPStatusCode.OK)
-            .json({
-                message: ErrorMessages.GETDATA,
-                data: getAllCourse
-            })
     } catch (error) {
         return res
         .status(HTTPStatusCode.INTERNAL_SERVER)
@@ -112,49 +138,54 @@ router.get('/getAllCourse', async(req,res) =>{
 })
 
 //Get Single Course
-router.get('/course/:id', async(req,res) =>{
+router.get('/course/:id',authenticateToken, async(req,res) =>{
     const id = req.params.id
+    const userid = req.user.id;
+    console.log(userid);
     try {
-        if(ObjectId.isValid(id)){
-            const getSingleCourse = await Course.findById({_id:id})
-            if(getSingleCourse){
-                    const category = await Category.findById({_id:getSingleCourse.category})
-                    if(category){
-                        getSingleCourse.category = category
-                    }
-                    const subCategory = await Subcategory.findById({_id:getSingleCourse.subCategory})
-                    if(subCategory){
-                        getSingleCourse.subCategory = subCategory
-                    }
-                    const programmingLanguage = await ProgrammingLanguage.findById({_id:getSingleCourse.programmingLanguage})
-                    if(programmingLanguage){
-                        getSingleCourse.programmingLanguage = programmingLanguage
-                    }
-                    const language = await Language.findById({_id:getSingleCourse.language});
-                    if(language){
-                        getSingleCourse.language = language
-                    }
+        const instructorExist = await Instructor.findById({_id:userid})
+        if(instructorExist){
+            if(ObjectId.isValid(id)){
+                const getSingleCourse = await Course.findOne({createdBy: userid})
+                if(getSingleCourse){
+                        const category = await Category.findById({_id:getSingleCourse.category})
+                        if(category){
+                            getSingleCourse.category = category
+                        }
+                        const subCategory = await Subcategory.findById({_id:getSingleCourse.subCategory})
+                        if(subCategory){
+                            getSingleCourse.subCategory = subCategory
+                        }
+                        const programmingLanguage = await ProgrammingLanguage.findById({_id:getSingleCourse.programmingLanguage})
+                        if(programmingLanguage){
+                            getSingleCourse.programmingLanguage = programmingLanguage
+                        }
+                        const language = await Language.findById({_id:getSingleCourse.language});
+                        if(language){
+                            getSingleCourse.language = language
+                        }
+                    return res
+                        .status(HTTPStatusCode.OK)
+                        .json({
+                            message: ErrorMessages.GETDATA,
+                            data: getSingleCourse
+                        })
+                }
+                else{
+                    return res
+                        .status(HTTPStatusCode.BAD_REQUEST)
+                        .json({
+                            message: ErrorMessages.NOT_EXISTS
+                        }) 
+                }
+            } else {
                 return res
-                    .status(HTTPStatusCode.OK)
-                    .json({
-                        message: ErrorMessages.GETDATA,
-                        data: getSingleCourse
-                    })
+                        .status(HTTPStatusCode.INTERNAL_SERVER)
+                        .json({
+                            message:ErrorMessages.WRONG_CREDENTIALS,
+                        
+                        })
             }
-            else{
-                return res
-                    .status(HTTPStatusCode.BAD_REQUEST)
-                    .json({
-                        message: ErrorMessages.NOT_EXISTS
-                    }) 
-            }
-        } else {
-            return res
-                    .status(HTTPStatusCode.INTERNAL_SERVER)
-                    .json({
-                        message:ErrorMessages.WRONG_CREDENTIALS,
-                     
-                    })
         }
     } catch (error) {
         return res
@@ -166,34 +197,52 @@ router.get('/course/:id', async(req,res) =>{
     }
 })
 //Update Course
-router.patch('/course/update/:id', async(req,res) =>{
-    const id = req.params.id
+router.patch('/course/update/:id',authenticateToken, async(req,res) =>{
+    const id = req.params.id;
+    console.log(req.body,"idddddd")
+    const userid = req.user.id;
+    console.log(userid);
     try {
-        if(ObjectId.isValid(id)){
-            const courseUpdate = await Course.findByIdAndUpdate(id, req.body, {
-                new:true
-            })
-            if(courseUpdate){
-                return res
-                    .status(HTTPStatusCode.OK)
-                    .json({
-                        message: ErrorMessages.UPDATED,
-                        data: courseUpdate
+        const instructorExist = await Instructor.findById({_id:userid})
+        if(instructorExist){
+            if(ObjectId.isValid(id)){
+                console.log(instructorExist);
+                // if(getSingleCourse){
+                    const courseUpdate = await Course.findByIdAndUpdate(id,req.body, {
+                        new:true
                     })
-            }
-            else{
-                return res
-                    .status(HTTPStatusCode.BAD_REQUEST)
-                    .json({
-                        message: ErrorMessages.NOT_EXISTS
-                    }) 
-            }
-        } else {
-            return res
-                    .status(HTTPStatusCode.INTERNAL_SERVER)
-                    .json({
-                        message:ErrorMessages.WRONG_CREDENTIALS,
-                    })
+                    console.log("cu",courseUpdate);
+                    if(courseUpdate){
+                        return res
+                            .status(HTTPStatusCode.OK)
+                            .json({
+                                message: ErrorMessages.UPDATED,
+                                data: courseUpdate
+                            })
+                    }
+                    else{
+                        return res
+                            .status(HTTPStatusCode.BAD_REQUEST)
+                            .json({
+                                message: ErrorMessages.NOT_EXISTS
+                            }) 
+                    }
+                }
+                else{
+                    return res
+                        .status(HTTPStatusCode.BAD_REQUEST)
+                        .json({
+                            message: ErrorMessages.NOT_EXISTS
+                        }) 
+                }
+              
+            // } else {
+            //     return res
+            //             .status(HTTPStatusCode.INTERNAL_SERVER)
+            //             .json({
+            //                 message:ErrorMessages.WRONG_CREDENTIALS,
+            //             })
+            // }
         }
     } catch (error) {
         return res
@@ -205,33 +254,38 @@ router.patch('/course/update/:id', async(req,res) =>{
     }
 })
 //Delete Course
-router.delete('/course/delete/:id', async(req,res) =>{
+router.delete('/course/delete/:id',authenticateToken, async(req,res) =>{
     const id = req.params.id
+    const userid = req.user.id;
+    console.log(userid);
     try {
-        if(ObjectId.isValid(id)){
-            const courseDelete = await Course.findByIdAndDelete({_id:id});
-            if(courseDelete){
+        const instructorExist = await Instructor.findById({_id:userid})
+        if(instructorExist){
+            if(ObjectId.isValid(id)){
+                const courseDelete = await Course.findByIdAndDelete({_id:id});
+                if(courseDelete){
+                    return res
+                        .status(HTTPStatusCode.OK)
+                        .json({
+                            message: ErrorMessages.DELETED,
+                            data: courseDelete
+                        })
+                }
+                else{
+                    return res
+                        .status(HTTPStatusCode.BAD_REQUEST)
+                        .json({
+                            message: ErrorMessages.NOT_EXISTS
+                        }) 
+                }
+            } else {
                 return res
-                    .status(HTTPStatusCode.OK)
-                    .json({
-                        message: ErrorMessages.DELETED,
-                        data: courseDelete
-                    })
+                        .status(HTTPStatusCode.INTERNAL_SERVER)
+                        .json({
+                            message:ErrorMessages.WRONG_CREDENTIALS,
+
+                        })
             }
-            else{
-                return res
-                    .status(HTTPStatusCode.BAD_REQUEST)
-                    .json({
-                        message: ErrorMessages.NOT_EXISTS
-                    }) 
-            }
-        } else {
-            return res
-                    .status(HTTPStatusCode.INTERNAL_SERVER)
-                    .json({
-                        message:ErrorMessages.WRONG_CREDENTIALS,
-                        
-                    })
         }
     } catch (error) {
         return res
